fix plot layer
