+++ examples/plot/plot-layer/surface-layer.js
@@ -1,24 +1,21 @@
 import {Layer} from 'deck.gl';
 import {GL, Model, Geometry} from 'luma.gl';
 
-import {scaleLinear} from 'd3-scale';
 import surfaceVertex from './surface-vertex.glsl';
 import fragmentShader from './fragment.glsl';
 
 const DEFAULT_COLOR = [0, 0, 0, 255];
-const DEFAULT_GET_SCALE = () => scaleLinear();
 
 const defaultProps = {
   data: [],
   getPosition: () => [0, 0, 0],
   getColor: () => DEFAULT_COLOR,
-  getXScale: DEFAULT_GET_SCALE,
-  getYScale: DEFAULT_GET_SCALE,
-  getZScale: DEFAULT_GET_SCALE,
+  xScale: null,
+  yScale: null,
+  zScale: null,
   uCount: 100,
   vCount: 100,
-  lightStrength: 0.1,
-  onUpdate: () => {}
+  lightStrength: 0.1
 };
 
 /*
@@ -30,9 +27,9 @@ const defaultProps = {
  * @param {Function} [props.getPosition] - method called to get [x, y, z] from (u,v) values
  * @param {Function} [props.getColor] - method called to get color from (x,y,z)
       returns [r,g,b,a].
- * @param {Function} [props.getXScale] - returns a d3 scale from (params = {min, max})
- * @param {Function} [props.getYScale] - returns a d3 scale from (params = {min, max})
- * @param {Function} [props.getZScale] - returns a d3 scale from (params = {min, max})
+ * @param {d3.scale} [props.xScale] - a d3 scale for the x axis
+ * @param {d3.scale} [props.yScale] - a d3 scale for the y axis
+ * @param {d3.scale} [props.zScale] - a d3 scale for the z axis
  * @param {Integer} [props.uCount] - number of samples within x range
  * @param {Integer} [props.vCount] - number of samples within y range
  * @param {Number} [props.lightStrength] - front light strength
@@ -182,15 +179,7 @@ export default class SurfaceLayer extends Layer {
   /* eslint-disable max-statements */
   calculatePositions(attribute) {
     const {vertexCount} = this.state;
-    const {uCount, vCount, getPosition, getXScale, getYScale, getZScale} = this.props;
-
-    // calculate z range
-    let xMin = Infinity;
-    let xMax = -Infinity;
-    let yMin = Infinity;
-    let yMax = -Infinity;
-    let zMin = Infinity;
-    let zMax = -Infinity;
+    const {uCount, vCount, getPosition, xScale, yScale, zScale} = this.props;
 
     const value = new Float32Array(vertexCount * attribute.size);
 
@@ -199,51 +188,21 @@ export default class SurfaceLayer extends Layer {
       for (let uIndex = 0; uIndex < uCount; uIndex++) {
         const u = uIndex / (uCount - 1);
         const v = vIndex / (vCount - 1);
-        let [x, y, z] = getPosition(u, v);
+        const [x, y, z] = getPosition(u, v);
 
         const isXFinite = isFinite(x);
         const isYFinite = isFinite(y);
         const isZFinite = isFinite(z);
-        if (!isXFinite) {
-          x = 0;
-        }
-        if (!isYFinite) {
-          y = 0;
-        }
-        if (!isZFinite) {
-          z = 0;
-        }
-
-        xMin = Math.min(xMin, x);
-        xMax = Math.max(xMax, x);
-        yMin = Math.min(yMin, y);
-        yMax = Math.max(yMax, y);
-        zMin = Math.min(zMin, z);
-        zMax = Math.max(zMax, z);
 
         // swap z and y: y is up in the default viewport
-        value[i++] = x;
-        value[i++] = z;
-        value[i++] = y;
+        value[i++] = isXFinite ? xScale(x) : xScale.range()[0];
+        value[i++] = isZFinite ? zScale(z) : zScale.range()[0];
+        value[i++] = isYFinite ? yScale(y) : yScale.range()[0];
         value[i++] = isXFinite && isYFinite && isZFinite ? 0 : 1;
       }
     }
 
-    const xScale = getXScale({min: xMin, max: xMax});
-    const yScale = getYScale({min: yMin, max: yMax});
-    const zScale = getZScale({min: zMin, max: zMax});
-
-    for (let j = 0; j < vertexCount; j++) {
-      const startIndex = j * 4;
-      if (!value[startIndex + 3]) {
-        value[startIndex] = xScale(value[startIndex]);
-        value[startIndex + 1] = zScale(value[startIndex + 1]);
-        value[startIndex + 2] = yScale(value[startIndex + 2]);
-      }
-    }
-
     attribute.value = value;
-    this.props.onUpdate({xScale, yScale, zScale});
   }
   /* eslint-enable max-statements */
